add feature to show bike in start position when there's no rec
